Several cleanups.
- Added some ASSERTs to check buffer access. These should not happen, but,
if there's some bug, they will, so: ASSERT()
- We now compile w/out warnings on MSVC - replaced usages of all
"insecure" C standard library functions
- a small issue in Windows makefile was fixed

+++ lib/sockets/pbpal_adns_sockets.c
@@ -119,13 +119,13 @@ static unsigned char* dns_qname_encode(uint8_t *dns, size_t n, uint8_t const* ho
         }
         else {
             size_t d = dest - lpos;
-	    *dest++ = '\0';
+            *dest++ = '\0';
             if (d > 63) {
                 /* label too long */
                 return NULL;
             }
 
-            *lpos =  d - 1;
+            *lpos =  (uint8_t)(d - 1);
             lpos += d;
 
             if ('\0' == hc) {
